requirements.txt
appdirs==1.4.3
asgiref==3.2.3
attrs==19.3.0
black==19.10b0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
codecov==2.0.16
colorama==0.4.3
coverage==5.0.3
crayons==0.3.0
distlib==0.3.0
Django==3.0.3
django-csvimport==2.13
django-excel==0.0.10
django-filter==2.2.0
django-phone-field==1.8.0
django-widget-tweaks==1.4.5
djangorestframework==3.11.0
entrypoints==0.3
et-xmlfile==1.0.1
filelock==3.0.12
flake8==3.7.9
fs==2.4.11
idna==2.9
isort==4.3.21
jdcal==1.4.1
Jinja2==2.11.1
jinja2-fsloader==0.3.0
lml==0.0.9
lxml==4.5.0
MarkupSafe==1.1.1
mccabe==0.6.1
moban==0.7.0
nose==1.3.7
openpyxl==3.0.3
pathspec==0.7.0
pycodestyle==2.5.0
pyecharts-jupyter-installer==0.0.3
pyexcel==0.5.15
pyexcel-ezodf==0.3.4
pyexcel-handsontable==0.0.2
pyexcel-io==0.5.20
pyexcel-ods3==0.5.3
pyexcel-pygal==0.0.2
pyexcel-webio==0.1.4
pyexcel-xls==0.5.8
pyexcel-xlsx==0.5.8
pyflakes==2.1.1
pygal==2.4.0
pytz==2019.3
regex==2020.2.20
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.14.0
sqlparse==0.3.0
texttable==1.6.2
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.8
virtualenv==20.0.8
virtualenvwrapper-win==1.2.6
xlrd==1.2.0
xlwt==1.3.0
